# [FLINK-36989][runtime] Fix scheduler benchmark regression caused by ConsumedSubpartitionContext