--- a/flink-runtime/src/test/java/org/apache/flink/runtime/deployment/ConsumedSubpartitionContextTest.java
+++ b/flink-runtime/src/test/java/org/apache/flink/runtime/deployment/ConsumedSubpartitionContextTest.java
@@ -21,6 +21,7 @@ package org.apache.flink.runtime.deployment;
 import org.apache.flink.runtime.executiongraph.IndexRange;
 import org.apache.flink.runtime.jobgraph.IntermediateDataSetID;
 import org.apache.flink.runtime.jobgraph.IntermediateResultPartitionID;
+import org.apache.flink.runtime.scheduler.strategy.ConsumedPartitionGroup;
 
 import org.junit.jupiter.api.Test;
 
@@ -29,6 +30,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 
+import static org.apache.flink.runtime.io.network.partition.ResultPartitionType.BLOCKING;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /** Tests for {@link ConsumedSubpartitionContext}. */
@@ -40,17 +42,13 @@ class ConsumedSubpartitionContextTest {
                         new IndexRange(0, 1), new IndexRange(0, 2),
                         new IndexRange(2, 3), new IndexRange(3, 5));
 
-        List<IntermediateResultPartitionID> consumedPartitionIds = new ArrayList<>();
-
-        IntermediateResultPartitionID[] partitions = new IntermediateResultPartitionID[4];
-        for (int i = 0; i < partitions.length; i++) {
-            partitions[i] = new IntermediateResultPartitionID(new IntermediateDataSetID(), i);
-            consumedPartitionIds.add(partitions[i]);
-        }
+        List<IntermediateResultPartitionID> partitions = createPartitions();
+        ConsumedPartitionGroup consumedPartitionGroup =
+                ConsumedPartitionGroup.fromMultiplePartitions(4, partitions, BLOCKING);
 
         ConsumedSubpartitionContext context =
                 ConsumedSubpartitionContext.buildConsumedSubpartitionContext(
-                        consumedSubpartitionGroups, consumedPartitionIds.iterator(), partitions);
+                        consumedSubpartitionGroups, consumedPartitionGroup, partitions::get);
 
         assertThat(context.getNumConsumedShuffleDescriptors()).isEqualTo(4);
 
@@ -71,17 +69,13 @@ class ConsumedSubpartitionContextTest {
                         new IndexRange(3, 3), new IndexRange(1, 1),
                         new IndexRange(0, 0), new IndexRange(0, 1));
 
-        List<IntermediateResultPartitionID> consumedPartitionIds = new ArrayList<>();
-
-        IntermediateResultPartitionID[] partitions = new IntermediateResultPartitionID[4];
-        for (int i = 0; i < partitions.length; i++) {
-            partitions[i] = new IntermediateResultPartitionID(new IntermediateDataSetID(), i);
-            consumedPartitionIds.add(partitions[i]);
-        }
+        List<IntermediateResultPartitionID> partitions = createPartitions();
+        ConsumedPartitionGroup consumedPartitionGroup =
+                ConsumedPartitionGroup.fromMultiplePartitions(4, partitions, BLOCKING);
 
         ConsumedSubpartitionContext context =
                 ConsumedSubpartitionContext.buildConsumedSubpartitionContext(
-                        consumedSubpartitionGroups, consumedPartitionIds.iterator(), partitions);
+                        consumedSubpartitionGroups, consumedPartitionGroup, partitions::get);
 
         assertThat(context.getNumConsumedShuffleDescriptors()).isEqualTo(2);
 
@@ -100,17 +94,13 @@ class ConsumedSubpartitionContextTest {
                         new IndexRange(0, 3), new IndexRange(1, 1),
                         new IndexRange(0, 1), new IndexRange(2, 2));
 
-        List<IntermediateResultPartitionID> consumedPartitionIds = new ArrayList<>();
-
-        IntermediateResultPartitionID[] partitions = new IntermediateResultPartitionID[4];
-        for (int i = 0; i < partitions.length; i++) {
-            partitions[i] = new IntermediateResultPartitionID(new IntermediateDataSetID(), i);
-            consumedPartitionIds.add(partitions[i]);
-        }
+        List<IntermediateResultPartitionID> partitions = createPartitions();
+        ConsumedPartitionGroup consumedPartitionGroup =
+                ConsumedPartitionGroup.fromMultiplePartitions(4, partitions, BLOCKING);
 
         ConsumedSubpartitionContext context =
                 ConsumedSubpartitionContext.buildConsumedSubpartitionContext(
-                        consumedSubpartitionGroups, consumedPartitionIds.iterator(), partitions);
+                        consumedSubpartitionGroups, consumedPartitionGroup, partitions::get);
 
         assertThat(context.getNumConsumedShuffleDescriptors()).isEqualTo(4);
 
@@ -143,5 +133,14 @@ class ConsumedSubpartitionContextTest {
 
         IndexRange subpartitionRange = context.getConsumedSubpartitionRange(2);
         assertThat(subpartitionRange).isEqualTo(consumedSubpartitionRange);
+    }
+
+    private static List<IntermediateResultPartitionID> createPartitions() {
+        List<IntermediateResultPartitionID> partitions = new ArrayList<>();
+        IntermediateDataSetID intermediateDataSetID = new IntermediateDataSetID();
+        for (int i = 0; i < 4; i++) {
+            partitions.add(new IntermediateResultPartitionID(intermediateDataSetID, i));
+        }
+        return partitions;
     }
 }