@zhuzhurk 2025-01-06
@flinkbot run azure