@noorall 2025-01-06
> Thanks @noorall for fixing. LGTM. Could you post the results of the three regressed benchmark cases after the fix?

I tested it on my own computer, and here are the test results:

  | Baseline (ms/op) | Before Fix (ms/op) | After Fix (ms/op)
-- | -- | -- | --
SchedulingAndDeployingBenchmarkExecutor.startScheduling(Streaming) | 253.740 ± 38.368 | 666.177 ± 132.292 | 318.120 ± 77.141
DeployingDownstreamTasksInBatchJobBenchmarkExecutor.deployDownstreamTasks(Batch) | 40.937 ± 10.214 | 423.432 ± 103.273 | 52.442 ± 29.368
HandleGlobalFailureAndRestartAllTasksBenchmarkExecutor.handleGlobalFailureAndRestartAllTasks.STREAMING | 334.504 ± 65.792 | 615.319 ± 81.008 | 402.614 ± 73.744
DeployingTasksInStreamingJobBenchmarkExecutor.deployAllTasks.STREAMING | 81.408 ± 21.408 | 455.895 ± 146.687 | 93.725 ± 36.951